gpc function to squelch the bad data
increasingly clear to me that the best way to handle the 3DR is to just
use it via the xbox adaptive controller, as this also strips out the
extra 2 axis that are detected as right joystick.
Gotta figure out a way to handle these two axis that doesn't brush up
against the limit of only being allowed one remapper.