number of changes to gyro
ME. provided a script to find the true report rate and delta time of
gyro input. Armed with the real deltaTime value, it seems that titan is
already handling deltaTime internally and it shouldn't be implemented
anyway.
ME. also provided access to the gyro data's true int16 range. I'm not
sure how useful it is to use this, might revert the changes later
Allowed deadzones to reach 100%, useful for deadzone finder to peg
joystick at 100%
some info revisions