uJSM implemented 'blinkerSync()', misc changes
blinkerSync is my silly name for a function that blinks the controller led at a
given interval. The goal is to adjust the duty cycle in real time until
the blinking matchis up with the max stick game speed, then that
information can be used to suggest a max stick game speed value to the
user.

changed maxstickgamespeed var type to save on some memory.